🚀 feat(RefreshScheduler): Schedule background refresh tasks on app backgrounding
Previously, the `RefreshScheduler` only scheduled a background refresh task when the application became active. This commit modifies the `RefreshScheduler` to also schedule a background refresh task when the application enters the background.

This change ensures that the application can initiate background refreshes more reliably, even if the app is not brought back to the foreground within the usual 15-minute interval. The `applicationDidEnterBackground` method in `AppDelegate` has been removed as its functionality is now handled by listening to `UIApplication.didEnterBackgroundNotification` within the `RefreshScheduler`.

New functionality added:
- `startListeningToBackgrounding()` method to listen for the `didEnterBackgroundNotification`.
- `backgrounded()` method to schedule the background refresh task when the app enters the background.

The `AppDelegate` no longer needs to explicitly call `refreshScheduler.schedule(in: 15.minutes)` in `applicationDidEnterBackground` as this is now handled internally by the `RefreshScheduler`.