🐛 fix(migration): skip migration when testing
This commit skips the migration process when running tests in `MigrationTests.swift`. This is because the tests are designed to specifically test the migration logic, and applying a migration before the test runs can lead to unexpected behavior or test failures. By skipping the migration, the tests can start with a clean slate and accurately verify the migration process itself.