🚨 refactor(logging): Configure SentryLogHandler to be more granular
Update the `SentryLogHandler` to use a more appropriate default `logLevel` of `.trace`. This change allows for more detailed logging to be sent to Sentry when necessary, rather than being restricted to `.notice` and above.

Additionally, introduce a conditional check `guard AppInfo.myDevice || level > .notice else { return }`. This ensures that less critical logs (below `.notice`) are only sent when running on a device, preventing excessive logging on simulators or for internal testing. This enhances control over the volume and relevance of logs sent to Sentry.