🐔 chore(cache): Reduce cache purge cadence because 12 hours was clearly too generous
Apparently, someone thought a 12-hour purge cycle for the cache was a good idea, leaving bloat to fester for half a day. This commit bravely shortens the interval to a more sensible 4 hours. Future generations will thank us for not letting our storage fill up with digital detritus for an entire day.