✨ refactor(search): Initialize SearchViewModel directly in ContentView and SearchView
Removes the uncached FactoryKit injection for `SearchViewModel` in `ContentView` and `SearchView`. Instead, the `SearchViewModel` is now initialized directly within `SearchView` and passed as a parameter. This simplifies dependency management for the `SearchViewModel` as it does not need to be managed globally by FactoryKit when its lifecycle is tied to the `SearchView`'s lifecycle.

Additionally, the `trendingSections` property in `SearchViewModel` has been changed from an `IdentifiedArray` to a standard `Array` as the `id: \.title` parameter for `IdentifiedArray` was not being utilized. This change reduces the overhead associated with `IdentifiedArray` when its unique identification capabilities are not required.